Introduce nodejs-repl-eval-script defcustom
This allows users to prepend, replace or event disable the automatic
execution of -e ... at the process command line.
The option should be a cons cell where the car is a the Javascript
string to evaluate and the cdr is the behavior nodejs-repl should apply
while bootstrapping, e.g.:

 - (setq nodejs-repl-eval-script '(\"...\" . prepend))
   prepends the code to the current bootstrapping code

 - (setq nodejs-repl-eval-script '(\"...\" . replace))
   replaces the current bootstrapping code

By default, the number of columns of the process is set and the repl is
started. Note that:

 - (setq nodejs-repl-eval-script '(\"\" . replace)) ;; empty

Completely avoids passing the eval script to the Node.js process.